Hide the menu if there are no pages
The intention is that the menu should not be shown
if there are no pages, but there is bug in the Minima theme:

    {% if page_paths %}

does not work as empty array is true. Add additional check for empty array:

    {%- if page_paths and page_paths != empty -%}

Also some pages may have paths but not titles (e.g. the Atom feed).
Filter them as only pages with titles are added to the menu:

    where_exp: "page", "page.title"

The rest of header.html is just copy+paste from the Minima source.