Add include snippet extension
Including only a part of a file is quite useful when writing a blog,
as it allows you to include only the important part of the code
instead of the whole file. Unfortunately the built-in include tag
does not allow that, but it does allow you to override read_file.
Use that to allow the inclusion of code snippets.

Also usually the indentation of the first line is
not desired in snippets, so the new tags will strip it.